Move thread sources and headers into subdirectory
