+++ embedded/SUFST/Src/Threads/sensor_thread.c
